de-duplicate disassember (nw)
